example_pipeline/requirements.txt
dlt[filesystem]>=1.0.0
psycopg2-binary>=2.9.9